# docs/swagger/swagger.yaml
basePath: /
definitions:
  api.errorResponse:
    properties:
      code:
        $ref: '#/definitions/domain.Code'
      message:
        type: string
    type: object
  domain.Code:
    enum:
    - invalid_argument
    - not_found
    - already_exists
    - permission_denied
    - internal
    - unauthenticated
    type: string
    x-enum-varnames:
    - CodeInvalidArgument
    - CodeNotFound
    - CodeAlreadyExists
    - CodePermissionDenied
    - CodeInternal
    - CodeUnauthenticated
...
  dto.JokeLikesResponse:
    properties:
      likes_number:
        type: integer
    type: object
...
host: localhost:3000
info:
  contact:
    email: valeriy.tymofieiev@gmail.com
    name: API Support
    url: https://github.com/abc-valera
  description: API for Flugo social network
  license:
    name: MIT
    url: https://github.com/abc-valera
  title: Flugo
  version: "0.1"
paths:
...
  /likes/{joke_id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: authorization
        required: true
        type: string
      - description: ID of required joke
        in: path
        name: joke_id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.JokeLikesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Returns likes number for specified joke.
      tags:
      - likes
...
  /me/likes/{joke_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: authorization
        required: true
        type: string
      - description: ID of required joke
        in: path
        name: joke_id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Deletes like (made by current user) of a specified joke.
      tags:
      - my likes
    post:
      consumes:
      - application/json
      parameters:
      - description: access token
        in: header
        name: authorization
        required: true
        type: string
      - description: ID of required joke
        in: path
        name: joke_id
        required: true
        type: integer
      produces:
      - text/plain
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.errorResponse'
      summary: Create new like by current user for specified joke.
      tags:
      - my likes
...
swagger: "2.0"